+++ core/src/rest/stock/ownership/mod.rs
@@ -9,5 +9,4 @@ mod tdcc_distribution;
 pub use director_holdings::DirectorHoldingsRequestBuilder;
 pub use etf_holdings::EtfHoldingsRequestBuilder;
 pub use institutional_trades::InstitutionalTradesRequestBuilder;
-pub use range::HoldingsSort;
 pub use tdcc_distribution::TdccDistributionRequestBuilder;